# [FIX] mrp, stock: speedup MO FormView opening