@robodoo 2022-11-29
[Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/106799).